Addressed formatting suggestions following pull request. Added examples of paged responses to the readme. Changed function name to APIProvider.result(_:isPagedResponse:) and passed in the associated request to ensure the object types being tested can only be the type of response associated with the endpoint.
